Fix issue with items after 19th August being deleted
This bug was introduced when I added the previous commit. The previous commit
fixed the fact that directory was not specified when opening the work_log files.

Unfortunately, I only fixed the output file location to use the script's own
directory (instead of pwd when the script is called by crontab, which is ~).

This meant that whenever the script read in the existing items, it read from the
old file, then output to the new file location. The upshot is that new items
were only added to the output file. This meant that whenever they "fell off the
bottom" of the RSS feed, then they were not read in again because the script
read from the old work_log file. This file was frozen in time at 19th August
because the next day I fixed the script to use the correct output location (but
not the correct input location).

I have ended up with a few backups of the logs over the past month and a half so
I'll manually merge those items which were preserved that way.